+++ packages/platform-core/src/usage-data/get_usage_data_url.test.ts
@@ -8,7 +8,7 @@ void describe('getUrl', () => {
     delete require.cache[require.resolve('./get_usage_data_url')];
   });
   void test('that prod URL is returned when the env for beta URL is not set', () => {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    // eslint-disable-next-line @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires
     const { getUrl } = require('./get_usage_data_url');
     assert.equal(
       url.format(getUrl()),
@@ -19,7 +19,7 @@ void describe('getUrl', () => {
   void test('that BETA URL is returned when the env for beta URL is set', () => {
     process.env.AMPLIFY_BACKEND_USAGE_TRACKING_ENDPOINT =
       'https://aws.amazon.com/amplify/';
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    // eslint-disable-next-line @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires
     const { getUrl } = require('./get_usage_data_url');
     assert.equal(url.format(getUrl()), 'https://aws.amazon.com/amplify/');
   });